impl Display for Warning
