+++ edgedb-tokio/src/raw/mod.rs
@@ -13,7 +13,6 @@ use std::sync::{Arc, Mutex as BlockingMutex};
 use std::time::Duration;
 
 use bytes::{Bytes, BytesMut};
-use edgedb_protocol::annotations::Warning;
 use tls_api::TlsStream;
 use tokio::sync::{self, Semaphore};
 
@@ -186,25 +185,7 @@ impl<T> Response<T> {
 
     fn log_warnings(&self) {
         for w in &self.warnings {
-            let Warning {
-                filename,
-                line,
-                col,
-                r#type,
-                message,
-                ..
-            } = w;
-            let filename = filename
-                .as_ref()
-                .map(|f| format!("{f}:"))
-                .unwrap_or_default();
-            let line = line.clone().unwrap_or(1);
-            let col = col.clone().unwrap_or(1);
-
-            log::warn!(
-                target: "edgedb_tokio::warning",
-                "{type} at {filename}{line}:{col} {message}"
-            )
+            log::warn!(target: "edgedb_tokio::warning", "{w}");
         }
     }
 }
